backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
networkx
faiss-cpu
openai
python-dotenv
